Refactoring manejo de roles administrador y usuario

+++ controller/CategoriasController.php
@@ -25,22 +25,30 @@ class CategoriasController {
     }
     function deleteCategoria($id, $jugadores) {
         $this->helper->checkLoggedIn();  
-        if(empty($jugadores)){
-            $this->model->deleteCategoria($id);
-            header("Location: " . BASE_URL . 'categorias');
+        if($_SESSION['ADMINISTRADOR']==1){
+            if(empty($jugadores)){
+                $this->model->deleteCategoria($id);
+                header("Location: " . BASE_URL . 'categorias');
+            }else{
+                $categorias=$this->model->getCategorias();
+                $this->view->showCategorias($categorias,"Esta categoria no se puede eliminar");
+            }
         }else{
-            $categorias=$this->model->getCategorias();
-            $this->view->showCategorias($categorias,"Esta categoria no se puede eliminar");
+            header('Location:' . BASE_URL . 'categorias');
         }
     }
     function insertCategoria(){
         $this->helper->checkLoggedIn();
-        $nombreDeporte = $_REQUEST['nombreDeporte'];
-        $descripcion = $_REQUEST['descripcion'];        
-        $tipo_competencia = $_REQUEST['tipo_competencia'];
+        if($_SESSION['ADMINISTRADOR']==1){
+            $nombreDeporte = $_REQUEST['nombreDeporte'];
+            $descripcion = $_REQUEST['descripcion'];        
+            $tipo_competencia = $_REQUEST['tipo_competencia'];
 
-        $this->model->insertCategoria($nombreDeporte,$descripcion,$tipo_competencia);
-        header("Location: " . BASE_URL . 'categorias');
+            $this->model->insertCategoria($nombreDeporte,$descripcion,$tipo_competencia);
+            header("Location: " . BASE_URL . 'categorias');
+        }else{
+            header('Location:' . BASE_URL . 'categorias');
+        }   
     }
     function getCategoriaById($id){
         $this->helper->checkLoggedIn();
@@ -48,15 +56,17 @@ class CategoriasController {
         $this->view->showUpdateCategoria($categoria);
     }
     function updateCategoria($id){
-        $this->helper->checkLoggedIn();
-        $this->view->showUpdateCategoria($id);
-        $nombreDeporte = $_REQUEST['nombreDeporte'];
-        $descripcion = $_REQUEST['descripcion'];        
-        $tipo_competencia = $_REQUEST['tipo_competencia'];
+        if($_SESSION['ADMINISTRADOR']==1){
+            $this->helper->checkLoggedIn();
+            $this->view->showUpdateCategoria($id);
+            $nombreDeporte = $_REQUEST['nombreDeporte'];
+            $descripcion = $_REQUEST['descripcion'];        
+            $tipo_competencia = $_REQUEST['tipo_competencia'];
 
-        $this->model->updateCategoria($nombreDeporte,$descripcion,$tipo_competencia,$id);
-        header("Location: " . BASE_URL . 'categorias');
+            $this->model->updateCategoria($nombreDeporte,$descripcion,$tipo_competencia,$id);
+            header("Location: " . BASE_URL . 'categorias');
+        }else{
+            header('Location:' . BASE_URL . 'categorias');
+        }
     }
 }
-
-?>
